ajout et test de Nos produits par catégorie

+++ controleurs/c_gestionPanier.php
@@ -27,6 +27,19 @@ switch($action)
 		include("vues/v_panier.php");
 		break;
 	}
+	case 'viderPanier':
+	{
+		$n= nbProduitsDuPanier();
+		if($n >0){
+			$message = "panier vidé";
+			supprimerPanier();
+			include ("vues/v_message.php");
+		} else {
+			$message = "panier vide !!";
+			include ("vues/v_message.php");
+		}
+		break;
+	}
 	case 'passerCommande' :
 	    $n= nbProduitsDuPanier();
 		if($n>0)
@@ -53,9 +66,10 @@ switch($action)
 		else
 		{
 			$lesIdProduit = getLesIdProduitsDuPanier();
-			creerCommande($nom,$rue,$cp,$ville,$mail, $lesIdProduit );
-			$message = "Commande enregistrée";
-			supprimerPanier();
+			if ( creerCommande($nom,$rue,$cp,$ville,$mail, $lesIdProduit ) ){
+				$message = "Commande enregistrée";
+				supprimerPanier();
+			}
 			include ("vues/v_message.php");
 		}
 		break;
